refactor: domain 코드 이동, ktlintformat
