+++ app/src/main/java/woowacourse/shopping/view/cart/CartPresenter.kt
@@ -1,5 +1,6 @@
 package woowacourse.shopping.view.cart
 
+import woowacourse.shopping.domain.CartPagination
 import woowacourse.shopping.domain.CartProduct
 import woowacourse.shopping.domain.CartRepository
 import woowacourse.shopping.domain.ProductRepository
