+++ app/src/main/res/layout/item_cart_pagination.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="status"
-            type="woowacourse.shopping.model.CartPageStatus" />
+            type="woowacourse.shopping.domain.CartPageStatus" />
 
         <variable
             name="onItemClick"
