+++ domain/src/main/java/woowacourse/shopping/domain/PrevPagination.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.model
+package woowacourse.shopping.domain
 
 interface PrevPagination<T> {
     val isPrevEnabled: Boolean
